+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+### [v2.0.0] - 2017-06-13
+
+- Now the main interface is a function instead of class
+- toHex() function was changed to hex()
+- toRGB() function was changed to rgb()
+- New function channels() to return object representation
+- New function isValid() to check if input color is valid or not
+
 ### [v1.2.1] - 2017-06-11
 
 - Documentation update